// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const Role = {
  EMPLOYEE: 'EMPLOYEE',
  HR: 'HR',
  IT_SUPPORT: 'IT_SUPPORT',
  IT_ADMIN: 'IT_ADMIN',
} as const;

export type Role = (typeof Role)[keyof typeof Role];

export const TicketStatus = {
  OPEN: 'OPEN',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
} as const;

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketPriority = {
  CRITICAL: 'CRITICAL',
  HIGH: 'HIGH',
  LOW: 'LOW',
} as const;

export type TicketPriority =
  (typeof TicketPriority)[keyof typeof TicketPriority];

export const Department = {
  HR: 'HR',
  IT: 'IT',
} as const;

export type Department = (typeof Department)[keyof typeof Department];

export const AssetStatus = {
  AVAILABLE: 'AVAILABLE',
  ASSIGNED: 'ASSIGNED',
  MAINTENANCE: 'MAINTENANCE',
  RETIRED: 'RETIRED',
} as const;

export type AssetStatus = (typeof AssetStatus)[keyof typeof AssetStatus];

export const AssetType = {
  NETWORK: 'NETWORK',
  SOFTWARE: 'SOFTWARE',
  HARDWARE: 'HARDWARE',
} as const;

export type AssetType = (typeof AssetType)[keyof typeof AssetType];